refactor :: 불필요한 주석, 메서드 삭제. HttpSession 생성자 주입으로 변경, 기타 코드 리뷰와 리팩터링
