+++ src/main/resources/templates/recoverBookForm.html
@@ -5,7 +5,7 @@
 <body>
 <div layout:fragment="content">
   <div style="display: flex; justify-content: center;">
-    <form th:action="@{/books/api/v1/recovers}" th:object="${recoverBookRequest}" method="post" style="width: 50%; margin-top: 100px;">
+    <form th:action="@{/books/recovers}" th:object="${recoverBookRequest}" method="post" style="width: 50%; margin-top: 100px;">
       <label for="bookId">내가 등록한 책 목록</label>
       <select id="bookId" name="bookId" th:field="*{bookId}">
         <option value="0" disabled>회수할 책을 선택하세요</option>
